+++ genmeta-curl/src/main.rs
@@ -1,19 +1,11 @@
 use clap::Parser;
+use genmeta_curl::{Options, run};
+use snafu::Whatever;
 
 #[tokio::main]
-async fn main() {
-    tracing_subscriber::fmt()
-        .with_env_filter(
-            tracing_subscriber::EnvFilter::builder()
-                .with_default_directive(tracing_subscriber::filter::LevelFilter::OFF.into())
-                .from_env_lossy(),
-        )
-        .with_writer(std::io::stderr)
-        .init();
-
-    if let Err(error) = genmeta_curl::run(genmeta_curl::Options::parse()).await {
-        eprintln!("{error}");
-        tracing::error!("Exit with error: {}", error);
-        std::process::exit(1);
-    }
+#[snafu::report]
+async fn main() -> Result<(), Whatever> {
+    run(Options::parse()).await.inspect_err(|error| {
+        tracing::debug!(?error, "Exit with error");
+    })
 }