+++ genmeta-doctor/src/lib.rs
@@ -1,3 +1,5 @@
+use snafu::Whatever;
+
 #[derive(Debug, Clone, clap::Parser)]
 #[command(name = "doctor", version, about)]
 pub enum Options {
@@ -5,7 +7,7 @@ pub enum Options {
     Net(genmeta_nat::Options),
 }
 
-type Error = Box<dyn core::error::Error + Send + Sync>;
+type Error = Whatever;
 
 pub async fn run(options: Options) -> Result<(), Error> {
     match options {