+++ ssh-config/src/error.rs
@@ -7,13 +7,13 @@ use peg::{error::ParseError, str::LineCol};
 pub enum ParseConfigError {
     #[snafu(display("Too many values at {location}"))]
     TooManyArguments { location: LineCol },
-    #[snafu(display("Cannot read file at `{}`: {source}", path.display()))]
+    #[snafu(display("Cannot read file at `{}`", path.display()))]
     ReadAsset {
         path: PathBuf,
         source: std::io::Error,
     },
     #[cfg(feature = "openssh")]
-    #[snafu(display("Cannot parse URI at {location}: {source}"))]
+    #[snafu(display("Cannot parse URI at {location}"))]
     ParseUri {
         location: LineCol,
         source: <http::Uri as std::str::FromStr>::Err,
@@ -32,17 +32,17 @@ pub enum ReadConfigError {
     LocateConfig {
         message: String,
     },
-    #[snafu(display("Cannot read config file at `{}`: {source}", path.display()))]
+    #[snafu(display("Cannot read config file at `{}`", path.display()))]
     ReadConfig {
         path: PathBuf,
         source: io::Error,
     },
-    #[snafu(display("Cannot parse config file `{}`: {source}", path.display()))]
+    #[snafu(display("Cannot parse config file `{}`", path.display()))]
     LexConfig {
         path: PathBuf,
         source: ParseError<LineCol>,
     },
-    #[snafu(display("Cannot parse config file `{}`: {source}", path.display()))]
+    #[snafu(display("Cannot parse config file `{}`", path.display()))]
     ParseConfig {
         path: PathBuf,
         source: ParseConfigError,